[CALCITE-3584] Propagate hints when decorrelating a query
* Add a new interface Hintable#withHints(List) to set up the rel node
hints;
* Removes Hintable#mergeHints(List) because it is only a tool method and
i believe that rare cases would need to customize that merging rule;
* Move RelHintPropagateShuttle from SqlToRelConverter to RelOptUtil;
* Add a new method RelOptUtil#propagateRelHints(RelNode, boolean) to
support hints for sql-to-rel conversion and decorrelation;
* Copy the hints for flattened projection in RelStructuredTypeFlattener;
* Add test cases for hints propagation with decorrelation.
