+++ core/src/main/java/org/apache/calcite/rel/hint/Hintable.java
@@ -51,40 +51,47 @@ import java.util.Set;
 public interface Hintable {
 
   /**
-   * Attach list of hints to this relational expression, should be overridden by
+   * Attaches list of hints to this relational expression, should be overridden by
    * every logical node that supports hint. This method is only for
    * internal use during sql-to-rel conversion.
    *
-   * <p>The sub-class should return a new copy of the relational expression. We make
-   * the default implementation return the relational expression directly only
-   * because not every kind of relational expression supports hints.
+   * <p>The sub-class should return a new copy of the relational expression.
+   *
+   * <p>The default implementation merges the given hints with existing ones,
+   * put them in one list and eliminate the duplicates; then
+   * returns a new copy of this relational expression with the merged hints.
    *
    * @param hintList The hints to attach to this relational expression
    * @return Relational expression with the hints {@code hintList} attached
    */
   default RelNode attachHints(List<RelHint> hintList) {
+    Objects.requireNonNull(hintList);
+    final Set<RelHint> hints = new LinkedHashSet<>(getHints());
+    hints.addAll(hintList);
+    return withHints(new ArrayList<>(hints));
+  }
+
+  /**
+   * Returns a new relation expression with the specified hints {@code hintList}.
+   *
+   * <p>This method should be overridden by every logical node that supports hint.
+   * It is only for internal use during decorrelation.
+   *
+   * <p>The sub-class should return a new copy of the relational expression.
+   *
+   * <p>We make the default implementation return the relational expression directly
+   * only because not every kind of relational expression supports hints.
+   *
+   * @return Relational expression with set up hints
+   */
+  default RelNode withHints(List<RelHint> hintList) {
     return (RelNode) this;
   }
 
   /**
-   * @return The hints list of this relational expressions
+   * Returns the hints of this relational expressions as a list.
    */
   ImmutableList<RelHint> getHints();
-
-  /**
-   * Merge this relation expression's hints with the given hint list.
-   *
-   * <p>The default behavior is to put them in one list and eliminate the duplicates.
-   *
-   * @param hintList Hints to be merged
-   * @return A merged hint list
-   */
-  default List<RelHint> mergeHints(List<RelHint> hintList) {
-    Objects.requireNonNull(hintList);
-    final Set<RelHint> hints = new LinkedHashSet<>(getHints());
-    hints.addAll(hintList);
-    return new ArrayList<>(hints);
-  }
 }
 
 // End Hintable.java
