+++ core/src/main/java/org/apache/calcite/rel/logical/LogicalJoin.java
@@ -214,8 +214,8 @@ public final class LogicalJoin extends Join {
     return systemFieldList;
   }
 
-  @Override public RelNode attachHints(List<RelHint> hintList) {
-    return new LogicalJoin(getCluster(), traitSet, mergeHints(hintList),
+  @Override public RelNode withHints(List<RelHint> hintList) {
+    return new LogicalJoin(getCluster(), traitSet, hintList,
         left, right, condition, variablesSet, joinType, semiJoinDone, systemFieldList);
   }
 }
