+++ core/src/main/java/org/apache/calcite/rel/logical/LogicalTableScan.java
@@ -129,8 +129,8 @@ public final class LogicalTableScan extends TableScan {
     return create(cluster, relOptTable, new ArrayList<>());
   }
 
-  @Override public RelNode attachHints(List<RelHint> hintList) {
-    return new LogicalTableScan(getCluster(), traitSet, mergeHints(hintList), table);
+  @Override public RelNode withHints(List<RelHint> hintList) {
+    return new LogicalTableScan(getCluster(), traitSet, hintList, table);
   }
 }
