+++ core/src/main/java/org/apache/calcite/sql2rel/RelStructuredTypeFlattener.java
@@ -216,6 +216,7 @@ public class RelStructuredTypeFlattener implements ReflectiveVisitor {
       RelNode restructured = relBuilder.push(flattened)
           .projectNamed(structuringExps, resultFieldNames, true)
           .build();
+      restructured = RelOptUtil.copyRelHints(flattened, restructured);
       // REVIEW jvs 23-Mar-2005:  How do we make sure that this
       // implementation stays in Java?  Fennel can't handle
       // structured types.
@@ -731,8 +732,9 @@ public class RelStructuredTypeFlattener implements ReflectiveVisitor {
     RelNode newRel = rel.getTable().toRel(toRelContext);
     if (!SqlTypeUtil.isFlat(rel.getRowType())) {
       newRel = coverNewRelByFlatteningProjection(rel, newRel);
+    } else {
+      newRel = RelOptUtil.copyRelHints(rel, newRel);
     }
-    newRel = RelOptUtil.copyRelHints(rel, newRel);
     setNewForOldRel(rel, newRel);
   }
 
@@ -747,6 +749,7 @@ public class RelStructuredTypeFlattener implements ReflectiveVisitor {
     newRel = relBuilder.push(newRel)
         .projectNamed(projects, fieldNames, true)
         .build();
+    newRel = RelOptUtil.copyRelHints(rel, newRel);
     return newRel;
   }
