+++ core/src/main/java/org/apache/calcite/sql2rel/SqlToRelConverter.java
@@ -194,12 +194,10 @@ import java.util.ArrayList;
 import java.util.BitSet;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.Deque;
 import java.util.EnumSet;
 import java.util.HashMap;
 import java.util.HashSet;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -598,7 +596,7 @@ public class SqlToRelConverter {
       }
     }
     // propagate the hints.
-    result = result.accept(new RelHintPropagateShuttle(this.hintStrategies));
+    result = RelOptUtil.propagateRelHints(result, false);
     return RelRoot.of(result, validatedRowType, query.getKind())
         .withCollation(collation)
         .withHints(hints);
@@ -5637,126 +5635,6 @@ public class SqlToRelConverter {
       this.id = id;
       this.requiredColumns = requiredColumns;
       this.r = r;
-    }
-  }
-
-  /**
-   * A {@code RelShuttle} which propagates all the hints of relational expression to
-   * their children nodes.
-   *
-   * <p>Given a plan:
-   *
-   * <blockquote><pre>
-   *            Filter (Hint1)
-   *                |
-   *               Join
-   *              /    \
-   *            Scan  Project (Hint2)
-   *                     |
-   *                    Scan2
-   * </pre></blockquote>
-   *
-   * <p>Every hint has a {@code inheritPath} (integers list) which records its propagate path,
-   * number `0` represents the hint is propagated from the first(left) child,
-   * number `1` represents the hint is propagated from the second(right) child,
-   * so the plan would have hints path as follows
-   * (assumes each hint can be propagated to all child nodes):
-   *
-   * <ul>
-   *   <li>Filter would have hints {Hint1[]}</li>
-   *   <li>Join would have hints {Hint1[0]}</li>
-   *   <li>Scan would have hints {Hint1[0, 0]}</li>
-   *   <li>Project would have hints {Hint1[0,1], Hint2[]}</li>
-   *   <li>Scan2 would have hints {[Hint1[0, 1, 0], Hint2[0]}</li>
-   * </ul>
-   */
-  private static class RelHintPropagateShuttle extends RelShuttleImpl {
-    /**
-     * Stack recording the hints and its current inheritPath.
-     */
-    private final Deque<Pair<List<RelHint>, Deque<Integer>>> inheritPaths =
-        new ArrayDeque<>();
-
-    /**
-     * The hint strategies to decide if a hint should be attached to
-     * a relational expression.
-     */
-    private final HintStrategyTable hintStrategies;
-
-    RelHintPropagateShuttle(HintStrategyTable hintStrategies) {
-      this.hintStrategies = hintStrategies;
-    }
-
-    /**
-     * Visits a particular child of a parent.
-     */
-    protected RelNode visitChild(RelNode parent, int i, RelNode child) {
-      inheritPaths.forEach(inheritPath -> inheritPath.right.push(i));
-      try {
-        RelNode child2 = child.accept(this);
-        if (child2 != child) {
-          final List<RelNode> newInputs = new ArrayList<>(parent.getInputs());
-          newInputs.set(i, child2);
-          return parent.copy(parent.getTraitSet(), newInputs);
-        }
-        return parent;
-      } finally {
-        inheritPaths.forEach(inheritPath -> inheritPath.right.pop());
-      }
-    }
-
-    public RelNode visit(LogicalJoin join) {
-      final LogicalJoin join1 = (LogicalJoin) super.visit(join);
-      return attachHints(join1);
-    }
-
-    public RelNode visit(LogicalProject project) {
-      final List<RelHint> topHints = project.getHints();
-      final boolean hasHints = topHints != null && topHints.size() > 0;
-      if (hasHints) {
-        inheritPaths.push(Pair.of(topHints, new ArrayDeque<>()));
-      }
-      final RelNode project1 = super.visit(project);
-      if (hasHints) {
-        inheritPaths.pop();
-      }
-      return attachHints(project1);
-    }
-
-    public RelNode visit(TableScan scan) {
-      TableScan scan1 = (TableScan) super.visit(scan);
-      return attachHints(scan1);
-    }
-
-    private RelNode attachHints(RelNode original) {
-      assert original instanceof Hintable;
-      if (inheritPaths.size() > 0) {
-        final List<RelHint> hints = inheritPaths.stream()
-            .sorted(Comparator.comparingInt(o -> o.right.size()))
-            .map(path -> copyWithInheritPath(path.left, path.right))
-            .reduce(new ArrayList<>(), (acc, hints1) -> {
-              acc.addAll(hints1);
-              return acc;
-            });
-        final List<RelHint> filteredHints = hintStrategies.apply(hints, original);
-        if (filteredHints.size() > 0) {
-          return ((Hintable) original).attachHints(filteredHints);
-        }
-      }
-      return original;
-    }
-
-    private static List<RelHint> copyWithInheritPath(List<RelHint> hints,
-        Deque<Integer> inheritPath) {
-      // Copy the Dequeue in reverse order.
-      final List<Integer> path = new ArrayList<>();
-      final Iterator<Integer> iterator = inheritPath.descendingIterator();
-      while (iterator.hasNext()) {
-        path.add(iterator.next());
-      }
-      return hints.stream()
-          .map(hint -> hint.copy(path))
-          .collect(Collectors.toList());
     }
   }
